requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.6
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-filter==22.1
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
django-select2==8.1.0
install==1.3.5
phonenumberslite==8.13.6
Pillow==9.4.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
